Move join paths helper to common library
