+++ test/unit/lib/path.test.js
@@ -1,4 +1,4 @@
-const { joinPaths } = require('~/src/modules/form/helpers.js')
+const { joinPaths } = require('~/src/lib/path.js')
 
 describe('#joinPaths', () => {
   context('when the first path has an extra forward slash at the end', () => {
